Move bottom navigation handling out of MainActivity
