+++ app/src/main/res/values/ids.xml
@@ -28,5 +28,6 @@
     <item name="bottom_navigation_addfeed" type="id"/>
     <item name="bottom_navigation_subscriptions" type="id"/>
     <item name="bottom_navigation_more" type="id"/>
+    <item name="bottom_navigation_customize" type="id"/>
     <item name="bottom_navigation_settings" type="id"/>
 </resources>
